Add files via upload
THIS IS A PROJECT DONE BY PRAAVEEN PRAKASH TIWARI IN THE GUIDAANCE OF TEZOS 7 - DAYS HACKATHON. In this Project, I have made a cryptobot which is named as "Bowler Smashing". All the codes are written by me on SmartPy Contracts and the process was fully completed when it goes deploy by Temple Wallet and this smart contract is visible openly. It was my first project and I have tried my best for it. It was so great learning moment in my life.